requirements.txt
fastapi
uvicorn[standard]
python-multipart

ultralytics
sahi
torch
torchvision
opencv-python-headless
matplotlib
numpy
Pillow